Added VHDL and TCL code
The VHDL code is a collection of "demos" that it is hoped the Scratch editor will be able to realise. This commit includes the use of an external submodule for the test bench code. The TCL code covers an interactive way of testing using a TCL/TK graphical user interface.