Scripts changes & initialising ROM from a file
Mainly batch file changes to try and stop Vivado splurging unwanted files in the directory where it was started. The RISC CPU ROM is now initialised from a file using std.textio, and the file name specified via a generic that is set from a configuration. The text file management is only partially implemented as we need the customasm.exe assembler to be included.